requirements.txt
numpy>=1.26.4
transformers>=4.51.3
lightning>=2.5.2
torch>=2.5.1
peft>=0.10.0
hydra-core>=1.3.2
liger-kernel